+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # WIP: 0.3.2
 
 * Add `DatapipeApp.api` subapp to mount extra APIs
+* Add `step status` command that shows status (number of idx's to process) for a
+  specific step
 
 # 0.3.1